Make Protocol.scala only contain Message, move "core plugin" events out
This should hopefully be less confusing for future developers.
